+++ apps/docs/components/ExampleUseDebounce.tsx
@@ -1,5 +1,8 @@
+'use client';
+
 import { useState } from 'react';
 import { useDebounce } from '@enterwell/react-hooks';
+import { Button } from '@mui/material';
 
 export function ExampleUseDebounce() {
     const [value, setValue] = useState(0);
@@ -11,7 +14,7 @@ export function ExampleUseDebounce() {
         <div>
             <div>Value: {value}</div>
             <div>Value debounced (1s): {valueDebounced}</div>
-            <button onClick={() => setValue((curr) => curr + 1)}>Increment</button>
+            <Button variant='outlined' onClick={() => setValue((curr) => curr + 1)}>Increment</Button>
         </div>
     );
 }