+++ packages/react-ui/ItemAccordion/ItemAccordion.tsx
@@ -1,0 +1,12 @@
+import MuiAccordion, { type AccordionProps } from '@mui/material/Accordion';
+
+/**
+ * Item accordion component.
+ * 
+ * @param props - The props of the component
+ * @returns The ItemAccordion component.
+ * @public
+ */
+export function ItemAccordion(props: AccordionProps) {
+  return <MuiAccordion disableGutters {...props} />
+}